+++ integrationstests/src/main/java/de/raysha/lib/jsimpleshell/CommandsWithoutAnnotation.java
@@ -1,0 +1,8 @@
+package de.raysha.lib.jsimpleshell;
+
+public class CommandsWithoutAnnotation {
+
+	public String hello(){
+		return "Hello World";
+	}
+}